Fixes #15262: Update relay component - split configuration.rs
