+++ relay/sources/relayd/benches/runlog.rs
@@ -3,7 +3,7 @@ use diesel::{self, prelude::*};
 use flate2::read::GzDecoder;
 use openssl::{stack::Stack, x509::X509};
 use relayd::{
-    configuration::DatabaseConfig,
+    configuration::main::DatabaseConfig,
     data::{report::QueryableReport, RunInfo, RunLog},
     input::signature,
     output::database::{schema::ruddersysevents::dsl::*, *},
