+++ relay/sources/relayd/src/output/database.rs
@@ -29,7 +29,7 @@
 // along with Rudder.  If not, see <http://www.gnu.org/licenses/>.
 
 use crate::{
-    configuration::DatabaseConfig, data::report::QueryableReport, data::RunLog, error::Error,
+    configuration::main::DatabaseConfig, data::report::QueryableReport, data::RunLog, error::Error,
 };
 use diesel::{
     insert_into,
