+++ relay/sources/relayd/tests/relay_api.rs
@@ -1,4 +1,4 @@
-use relayd::{configuration::CliConfiguration, init_logger, start};
+use relayd::{configuration::cli::CliConfiguration, init_logger, start};
 use reqwest;
 use std::{fs::read_to_string, thread, time};
